src/ios/backgroundvideo.m: Use AVCaptureSessionPresetHigh instead of AVCaptureSessionPreset1920x1080
Seems like HQ 1920x1080 and 1280x720 resolutions is not supported for
front camera in each Apple device. For back camera these existing setting should be fine in
most cases (better support for 1920x1080). So try to honor best settings
by setting AVCaptureSessionPresetHigh, hoepfully fallbacks to the best
settings which is available in concrete device related to the hardware.